refactor: move everything in root Python module
