+++ examples/z_pull.py
@@ -16,8 +16,6 @@ import time
 import argparse
 import json
 import zenoh
-from zenoh.handlers import RingChannel
-from zenoh.subscriber import Reliability
 
 # --- Command line argument parsing --- --- --- --- --- ---
 parser = argparse.ArgumentParser(prog="z_pull", description="zenoh pull example")
@@ -76,9 +74,7 @@ parser.add_argument(
 
 args = parser.parse_args()
 conf = (
-    zenoh.config.Config.from_file(args.config)
-    if args.config is not None
-    else zenoh.config.default()
+    zenoh.Config.from_file(args.config) if args.config is not None else zenoh.Config()
 )
 if args.mode is not None:
     conf.insert_json5("mode", json.dumps(args.mode))
@@ -102,7 +98,9 @@ def main():
     # Subscriber doesn't receive messages over the RingBuffer size.
     # The oldest message is overwritten by the latest one.
     sub = session.declare_subscriber(
-        key, reliability=Reliability.RELIABLE, handler=RingChannel(args.size)
+        key,
+        reliability=zenoh.Reliability.RELIABLE,
+        handler=zenoh.handlers.RingChannel(args.size),
     )
 
     print("Press CTRL-C to quit...")
